compress: merge registry into tools, remove dead API, simplify shell I/O
Deleted registry.rs — Tool trait, ToolResult, and ToolRegistry now live
in tools.rs alongside their implementations. Removed pub
dispatch_tool_results wrapper (only used by same-module tests).
Replaced manual stdout/stderr reading in run_shell_command with
wait_with_output.
