+++ rust/loopflow/src/agent/mod.rs
@@ -1,5 +1,4 @@
 pub mod anthropic;
 pub mod context;
-pub mod registry;
 pub mod tools;
 pub mod turn;
